+++ package.json
@@ -4,6 +4,8 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
+    "dev:local": "next dev TARGET=http://localhost:5186",
+    "dev:mseller": "next dev TARGET=https://portal-int-api.mseller.app",
     "build": "next build",
     "start": "next start",
     "export": "next export",
